// src/lib/api/client/types.gen.ts
export type ContentBlock = {
    type: 'paragraph';
    text: string;
} | {
    type: 'heading';
    level: 2 | 3;
    text: string;
} | {
    type: 'blockquote';
    text: string;
} | {
    type: 'unordered-list';
    items: Array<string>;
} | {
    type: 'ordered-list';
    items: Array<string>;
} | {
    type: 'image';
    src: string;
    alt: string;
    caption?: string;
};